requirements.txt
Django==2.0.7
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
factory-boy==2.11.1
requests==2.18.4
